Instrument code changes to reduce compiler/analyzer warnings.

+++ insts/std/PVOC/PVOC.cpp
@@ -202,7 +202,7 @@ int PVOC::init(double *p, int n_args)
 	I	 = (int)p[8];		/* interpolation factor */
 	P	 = p[9];			/* oscillator bank pitch factor */
 	Np	= (int)p[10];		/* linear prediction order */
-	_oscThreshold  = p[11];		/* synthesis threshhold */
+	_oscThreshold  = (float)p[11];		/* synthesis threshhold */
 
 #ifdef debug
 	printf("initial PVOC parameters:\n" );
@@ -340,7 +340,7 @@ int PVOC::doUpdate()
 		else
 			_NP = int(N2);
 	}
-	_oscThreshold  = update(11, _inputFrames, _currentInputFrame);		/* synthesis threshhold */
+	_oscThreshold  = (float)update(11, _inputFrames, _currentInputFrame);		/* synthesis threshhold */
 	
 	return 0;
 }
